Solve issue : acknowledgment at the entity level only working with entities with roles sender (#8097)
And some refactoring : split the AcknowledgeService
