+++ ui/main/src/app/services/acknowlegment/AcknowledgeService.spec.ts
@@ -1,4 +1,4 @@
-/* Copyright (c) 2021-2025, RTE (http://www.rte-france.com)
+/* Copyright (c) 2025, RTE (http://www.rte-france.com)
  * See AUTHORS.txt
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
@@ -7,67 +7,40 @@
  * This file is part of the OperatorFabric project.
  */
 
-import {AcknowledgeService} from './AcknowledgeService';
-import {
-    AcknowledgmentAllowedEnum,
-    ConsideredAcknowledgedForUserWhenEnum,
-    Process,
-    Response,
-    State
-} from '@ofServices/processes/model/Processes';
-import {getOneCard} from '@tests/helpers';
-import {Card} from 'app/model/Card';
+import {setUserPerimeter} from '@tests/helpers';
 import {UserWithPerimeters} from '@ofServices/users/model/UserWithPerimeters';
 import {User} from '@ofServices/users/model/User';
-import {RightEnum} from '@ofServices/perimeters/model/Perimeter';
 import {EntitiesService} from '@ofServices/entities/EntitiesService';
-import {ProcessesServerMock} from '@tests/mocks/processesServer.mock';
-import {UsersService} from '../users/UsersService';
-import {UsersServerMock} from '@tests/mocks/UsersServer.mock';
 import {EntitiesServerMock} from '@tests/mocks/entitiesServer.mock';
 import {Entity} from '@ofServices/entities/model/Entity';
-import {ServerResponse, ServerResponseStatus} from '../../server/ServerResponse';
-import {ProcessesService} from '../processes/ProcessesService';
-import {PermissionEnum} from '@ofServices/groups/model/PermissionEnum';
 import {RoleEnum} from '@ofServices/entities/model/RoleEnum';
+import {AcknowledgeServer} from './server/AcknowledgeServer';
+import {AcknowledgeService} from './AcknowledgeService';
+import {PermissionEnum} from '@ofServices/groups/model/PermissionEnum';
+
+class AcknowledgeServerMock implements AcknowledgeServer {
+    public entitiesAcksPosted: string[] = undefined;
+    public cardUidPosted: string = undefined;
+    public entitiesAcksDeleted: string[] = undefined;
+    public cardUidDeleted: string = undefined;
+
+    postUserAcknowledgement(cardUid: string, entitiesAcks: string[]): any {
+        this.entitiesAcksPosted = entitiesAcks;
+        this.cardUidPosted = cardUid;
+        return null;
+    }
+    deleteUserAcknowledgement(cardUid: string, entitiesAcks: string[]): any {
+        this.entitiesAcksDeleted = entitiesAcks;
+        this.cardUidDeleted = cardUid;
+        return null;
+    }
+}
 
 describe('AcknowledgeService testing ', () => {
-    let card: Card;
-    let userMemberOfEntity1: User, userMemberOfEntity2: User, userMemberOfEntity1AndEntity3: User;
-    let statesList: Map<string, State>;
-    let usersServerMock: UsersServerMock;
-
+    let acknowledgeServerMock: AcknowledgeServerMock;
     beforeEach(() => {
-        userMemberOfEntity1 = new User('userMemberOfEntity1', 'firstName', 'lastName', null, ['group1'], ['ENTITY1']);
-        userMemberOfEntity2 = new User('userMemberOfEntity2', 'firstName', 'lastName', null, ['group1'], ['ENTITY2']);
-        userMemberOfEntity1AndEntity3 = new User(
-            'userMemberOfEntity1AndEntity3',
-            'firstName',
-            null,
-            'lastName',
-            ['group1'],
-            ['ENTITY1', 'ENTITY3']
-        );
-
-        statesList = new Map<string, State>();
-
-        usersServerMock = new UsersServerMock();
-        usersServerMock.setResponseForCurrentUserWithPerimeter(
-            new ServerResponse(getUserMemberOfEntity1WithPerimeter(), ServerResponseStatus.OK, '')
-        );
-        UsersService.setUsersServer(usersServerMock);
-
-        const processesServerMock = new ProcessesServerMock();
-        processesServerMock.setResponseForProcessesDefinition(
-            new ServerResponse(getTestProcesses(), ServerResponseStatus.OK, '')
-        );
-        processesServerMock.setResponseForProcessesWithAllVersions(
-            new ServerResponse(getTestProcesses(), ServerResponseStatus.OK, '')
-        );
-        ProcessesService.setProcessServer(processesServerMock);
-        ProcessesService.loadAllProcessesWithLatestVersion().subscribe();
-        ProcessesService.loadAllProcessesWithAllVersions().subscribe();
-        UsersService.loadUserWithPerimetersData().subscribe();
+        acknowledgeServerMock = new AcknowledgeServerMock();
+        AcknowledgeService.setAcknowledgeServer(acknowledgeServerMock);
 
         const mockEntitiesServer = new EntitiesServerMock();
         mockEntitiesServer.setEntities([
@@ -77,950 +50,46 @@ describe('AcknowledgeService testing ', () => {
             new Entity('ENTITY_FR', 'ENTITY FR', '', [RoleEnum.CARD_SENDER], null, null)
         ]);
         EntitiesService.setEntitiesServer(mockEntitiesServer);
-
         EntitiesService.loadAllEntitiesData().subscribe();
-
-        card = getOneCard({
-            process: 'testProcess',
-            processVersion: '1',
-            state: 'testState',
-            entitiesAllowedToRespond: ['ENTITY1']
-        });
     });
 
-    function getOneProcess(processTemplate?: any): Process {
-        processTemplate = processTemplate ?? {};
-        const states = new Map();
+    it('should ack at the entity level', async () => {
+        const user = new User('user', 'firstName', 'lastName', null, ['group1'], ['ENTITY1', 'ENTITY2']);
+        const userWithPerimeters = new UserWithPerimeters(user, [], []);
+        await setUserPerimeter(userWithPerimeters);
 
-        states.set(1, new State('template1', ['style1', 'style2']));
-        states.set(2, new State('template1', ['style1', 'style2']));
-
-        return new Process(
-            processTemplate.id ?? 'testId',
-            processTemplate.version ?? 'testVersion',
-            processTemplate.name ?? 'testName',
-            processTemplate.locales ?? undefined,
-            processTemplate.states ?? states
-        );
-    }
-
-    function getTestProcesses(): Process[] {
-        const state1 = new State();
-        state1.acknowledgmentAllowed = AcknowledgmentAllowedEnum.ALWAYS;
-        state1.consideredAcknowledgedForUserWhen = ConsideredAcknowledgedForUserWhenEnum.USER_HAS_ACKNOWLEDGED;
-
-        const state2 = new State();
-        state2.acknowledgmentAllowed = AcknowledgmentAllowedEnum.ALWAYS;
-        state2.consideredAcknowledgedForUserWhen =
-            ConsideredAcknowledgedForUserWhenEnum.ALL_ENTITIES_OF_USER_HAVE_ACKNOWLEDGED;
-
-        const statesList = new Map();
-        statesList.set('state1', state1);
-        statesList.set('state2', state2);
-
-        const testProcess = new Process('testProcess', '1', null, null, statesList);
-        return [testProcess];
-    }
-
-    function getUserMemberOfEntity1WithPerimeter(): UserWithPerimeters {
-        return new UserWithPerimeters(userMemberOfEntity1, [
-            {
-                process: 'testProcess',
-                state: 'state1',
-                rights: RightEnum.ReceiveAndWrite,
-                filteringNotificationAllowed: true
-            },
-            {
-                process: 'testProcess',
-                state: 'state2',
-                rights: RightEnum.ReceiveAndWrite,
-                filteringNotificationAllowed: true
-            }
-        ]);
-    }
-
-    function getUserMemberOfEntity1AndEntity3WithPerimeter(): UserWithPerimeters {
-        return new UserWithPerimeters(userMemberOfEntity1AndEntity3, [
-            {
-                process: 'testProcess',
-                state: 'state1',
-                rights: RightEnum.ReceiveAndWrite,
-                filteringNotificationAllowed: true
-            },
-            {
-                process: 'testProcess',
-                state: 'state2',
-                rights: RightEnum.ReceiveAndWrite,
-                filteringNotificationAllowed: true
-            }
-        ]);
-    }
-
-    function getUserMemberOfEntity1WithPerimeterAndReadonly(): UserWithPerimeters {
-        return new UserWithPerimeters(
-            userMemberOfEntity1,
-            [
-                {
-                    process: 'testProcess',
-                    state: 'state1',
-                    rights: RightEnum.ReceiveAndWrite,
-                    filteringNotificationAllowed: true
-                },
-                {
-                    process: 'testProcess',
-                    state: 'state2',
-                    rights: RightEnum.ReceiveAndWrite,
-                    filteringNotificationAllowed: true
-                }
-            ],
-            [PermissionEnum.READONLY]
-        );
-    }
-
-    it('acknowledgmentAllowed of the state is not present , isAcknowledgmentAllowed() must return true (default value)', () => {
-        statesList.set('testState', new State(null, null, null, null));
-        const processDefinition = getOneProcess({id: 'testProcess', version: '1', states: statesList});
-        const userWithPerimeters = new UserWithPerimeters(userMemberOfEntity1, [
-            {process: 'testProcess', state: 'testState', rights: RightEnum.Receive, filteringNotificationAllowed: true}
-        ]);
-
-        const res = AcknowledgeService.isAcknowledgmentAllowed(userWithPerimeters, card, processDefinition);
-        expect(res).toBeTrue();
+        AcknowledgeService.postAcknowledgement('test');
+        expect(acknowledgeServerMock.cardUidPosted).toEqual('test');
+        expect(acknowledgeServerMock.entitiesAcksPosted).toEqual(['ENTITY1', 'ENTITY2']);
     });
 
-    it('process does not exist , isAcknowledgmentAllowed() must return true (default value)', () => {
-        const userWithPerimeters = new UserWithPerimeters(userMemberOfEntity1, [
-            {process: 'testProcess', state: 'testState', rights: RightEnum.Receive, filteringNotificationAllowed: true}
-        ]);
-        const res = AcknowledgeService.isAcknowledgmentAllowed(userWithPerimeters, card, null);
-        expect(res).toBeTrue();
+    it('should not ack at the entity level is user is readonly ', async () => {
+        const user = new User('user', 'firstName', 'lastName', null, ['group1'], ['ENTITY1']);
+        const userWithPerimeters = new UserWithPerimeters(user, [], [PermissionEnum.READONLY]);
+        await setUserPerimeter(userWithPerimeters);
+
+        AcknowledgeService.postAcknowledgement('test');
+        expect(acknowledgeServerMock.cardUidPosted).toEqual('test');
+        expect(acknowledgeServerMock.entitiesAcksPosted).toEqual([]);
     });
 
-    it('state does not exist , isAcknowledgmentAllowed() must return true (default value)', () => {
-        statesList.set('dummyState', new State(null, null, null, null));
-        const processDefinition = getOneProcess({id: 'testProcess', version: '1', states: statesList});
-        const userWithPerimeters = new UserWithPerimeters(userMemberOfEntity1, [
-            {process: 'testProcess', state: 'testState', rights: RightEnum.Receive, filteringNotificationAllowed: true}
-        ]);
+    it('should unack at the entity level', async () => {
+        const user = new User('user', 'firstName', 'lastName', null, ['group1'], ['ENTITY1', 'ENTITY2']);
+        const userWithPerimeters = new UserWithPerimeters(user, [], []);
+        await setUserPerimeter(userWithPerimeters);
 
-        const res = AcknowledgeService.isAcknowledgmentAllowed(userWithPerimeters, card, processDefinition);
-        expect(res).toBeTrue();
+        AcknowledgeService.deleteAcknowledgement('test');
+        expect(acknowledgeServerMock.cardUidDeleted).toEqual('test');
+        expect(acknowledgeServerMock.entitiesAcksDeleted).toEqual(['ENTITY1', 'ENTITY2']);
     });
 
-    it('acknowledgmentAllowed of the state is Never, isAcknowledgmentAllowed() must return false', () => {
-        statesList.set('testState', new State(null, null, null, AcknowledgmentAllowedEnum.NEVER));
-        const processDefinition = getOneProcess({id: 'testProcess', version: '1', states: statesList});
+    it('should not unack at the entity level is user is readonly ', async () => {
+        const user = new User('user', 'firstName', 'lastName', null, ['group1'], ['ENTITY1']);
+        const userWithPerimeters = new UserWithPerimeters(user, [], [PermissionEnum.READONLY]);
+        await setUserPerimeter(userWithPerimeters);
 
-        const userWithPerimeters = new UserWithPerimeters(userMemberOfEntity1, [
-            {process: 'testProcess', state: 'testState', rights: RightEnum.Receive, filteringNotificationAllowed: true}
-        ]);
-
-        const res = AcknowledgeService.isAcknowledgmentAllowed(userWithPerimeters, card, processDefinition);
-        expect(res).toBeFalse();
+        AcknowledgeService.deleteAcknowledgement('test');
+        expect(acknowledgeServerMock.cardUidDeleted).toEqual('test');
+        expect(acknowledgeServerMock.entitiesAcksDeleted).toEqual([]);
     });
-
-    it('acknowledgmentAllowed of the state is Always, isAcknowledgmentAllowed() must return true', () => {
-        statesList.set('testState', new State(null, null, null, AcknowledgmentAllowedEnum.ALWAYS));
-        const processDefinition = getOneProcess({id: 'testProcess', version: '1', states: statesList});
-        const userWithPerimeters = new UserWithPerimeters(userMemberOfEntity1, [
-            {process: 'testProcess', state: 'testState', rights: RightEnum.Receive, filteringNotificationAllowed: true}
-        ]);
-
-        const res = AcknowledgeService.isAcknowledgmentAllowed(userWithPerimeters, card, processDefinition);
-        expect(res).toBeTrue();
-    });
-
-    it(
-        'acknowledgmentAllowed of the state is OnlyWhenResponseDisabledForUser, ' +
-            'user cannot respond (user is a member of entity allowed to respond but user rights for the state of the response is Receive), ' +
-            'isAcknowledgmentAllowed() must return true',
-        () => {
-            statesList.set(
-                'testState',
-                new State(
-                    null,
-                    null,
-                    new Response(null, 'responseState'),
-                    AcknowledgmentAllowedEnum.ONLY_WHEN_RESPONSE_DISABLED_FOR_USER
-                )
-            );
-            const processDefinition = getOneProcess({id: 'testProcess', version: '1', states: statesList});
-            const userWithPerimeters = new UserWithPerimeters(userMemberOfEntity1, [
-                {
-                    process: 'testProcess',
-                    state: 'responseState',
-                    rights: RightEnum.Receive,
-                    filteringNotificationAllowed: true
-                }
-            ]);
-
-            const res = AcknowledgeService.isAcknowledgmentAllowed(userWithPerimeters, card, processDefinition);
-            expect(res).toBeTrue();
-        }
-    );
-
-    it(
-        'acknowledgmentAllowed of the state is OnlyWhenResponseDisabledForUser, ' +
-            'user can respond (user is a member of entity allowed to respond and user rights for the state of the response is Write), ' +
-            'isAcknowledgmentAllowed() must return false',
-        () => {
-            statesList.set(
-                'testState',
-                new State(
-                    null,
-                    null,
-                    new Response(null, 'responseState'),
-                    AcknowledgmentAllowedEnum.ONLY_WHEN_RESPONSE_DISABLED_FOR_USER
-                )
-            );
-            const processDefinition = getOneProcess({id: 'testProcess', version: '1', states: statesList});
-            const userWithPerimeters = new UserWithPerimeters(userMemberOfEntity1, [
-                {
-                    process: 'testProcess',
-                    state: 'responseState',
-                    rights: RightEnum.ReceiveAndWrite,
-                    filteringNotificationAllowed: true
-                }
-            ]);
-
-            const res = AcknowledgeService.isAcknowledgmentAllowed(userWithPerimeters, card, processDefinition);
-            expect(res).toBeFalse();
-        }
-    );
-
-    it(
-        'acknowledgmentAllowed of the state is OnlyWhenResponseDisabledForUser, ' +
-            'user can respond (user is a member of entity allowed to respond and user rights for the state of the response is ReceiveAndWrite), ' +
-            'isAcknowledgmentAllowed() must return false',
-        () => {
-            statesList.set(
-                'testState',
-                new State(
-                    null,
-                    null,
-                    new Response(null, 'responseState'),
-                    AcknowledgmentAllowedEnum.ONLY_WHEN_RESPONSE_DISABLED_FOR_USER
-                )
-            );
-            const processDefinition = getOneProcess({id: 'testProcess', version: '1', states: statesList});
-            const userWithPerimeters = new UserWithPerimeters(userMemberOfEntity1, [
-                {
-                    process: 'testProcess',
-                    state: 'responseState',
-                    rights: RightEnum.ReceiveAndWrite,
-                    filteringNotificationAllowed: true
-                }
-            ]);
-
-            const res = AcknowledgeService.isAcknowledgmentAllowed(userWithPerimeters, card, processDefinition);
-            expect(res).toBeFalse();
-        }
-    );
-
-    it(
-        'acknowledgmentAllowed of the state is OnlyWhenResponseDisabledForUser, ' +
-            'user cannot respond (user is not a member of entity allowed to respond and user rights for the state of the response is Receive), ' +
-            'isAcknowledgmentAllowed() must return true',
-        () => {
-            statesList.set(
-                'testState',
-                new State(
-                    null,
-                    null,
-                    new Response(null, 'responseState'),
-                    AcknowledgmentAllowedEnum.ONLY_WHEN_RESPONSE_DISABLED_FOR_USER
-                )
-            );
-            const processDefinition = getOneProcess({id: 'testProcess', version: '1', states: statesList});
-            const userWithPerimeters = new UserWithPerimeters(userMemberOfEntity2, [
-                {
-                    process: 'testProcess',
-                    state: 'responseState',
-                    rights: RightEnum.Receive,
-                    filteringNotificationAllowed: true
-                }
-            ]);
-
-            const res = AcknowledgeService.isAcknowledgmentAllowed(userWithPerimeters, card, processDefinition);
-            expect(res).toBeTrue();
-        }
-    );
-
-    it(
-        'acknowledgmentAllowed of the state is OnlyWhenResponseDisabledForUser, ' +
-            'user cannot respond (user is not a member of entity allowed to respond and user rights for the state of the response is Write), ' +
-            'isAcknowledgmentAllowed() must return true',
-        () => {
-            statesList.set(
-                'testState',
-                new State(
-                    null,
-                    null,
-                    new Response(null, 'responseState'),
-                    AcknowledgmentAllowedEnum.ONLY_WHEN_RESPONSE_DISABLED_FOR_USER
-                )
-            );
-            const processDefinition = getOneProcess({id: 'testProcess', version: '1', states: statesList});
-            const userWithPerimeters = new UserWithPerimeters(userMemberOfEntity2, [
-                {
-                    process: 'testProcess',
-                    state: 'responseState',
-                    rights: RightEnum.ReceiveAndWrite,
-                    filteringNotificationAllowed: true
-                }
-            ]);
-
-            const res = AcknowledgeService.isAcknowledgmentAllowed(userWithPerimeters, card, processDefinition);
-            expect(res).toBeTrue();
-        }
-    );
-
-    it(
-        'acknowledgmentAllowed of the state is OnlyWhenResponseDisabledForUser, ' +
-            'user cannot respond (user is not a member of entity allowed to respond and user rights for the state of the response is ReceiveAndWrite), ' +
-            'isAcknowledgmentAllowed() must return true',
-        () => {
-            statesList.set(
-                'testState',
-                new State(
-                    null,
-                    null,
-                    new Response(null, 'responseState'),
-                    AcknowledgmentAllowedEnum.ONLY_WHEN_RESPONSE_DISABLED_FOR_USER
-                )
-            );
-            const processDefinition = getOneProcess({id: 'testProcess', version: '1', states: statesList});
-            const userWithPerimeters = new UserWithPerimeters(userMemberOfEntity2, [
-                {
-                    process: 'testProcess',
-                    state: 'responseState',
-                    rights: RightEnum.ReceiveAndWrite,
-                    filteringNotificationAllowed: true
-                }
-            ]);
-
-            const res = AcknowledgeService.isAcknowledgmentAllowed(userWithPerimeters, card, processDefinition);
-            expect(res).toBeTrue();
-        }
-    );
-
-    it(
-        'acknowledgmentAllowed of the state is OnlyWhenResponseDisabledForUser, ' +
-            'user can respond (user is a member of entity allowed to respond (ENTITY_1) and user rights for the state of the response is Write), ' +
-            'isAcknowledgmentAllowed() must return false',
-        () => {
-            statesList.set(
-                'testState',
-                new State(
-                    null,
-                    null,
-                    new Response(null, 'responseState'),
-                    AcknowledgmentAllowedEnum.ONLY_WHEN_RESPONSE_DISABLED_FOR_USER
-                )
-            );
-            const processDefinition = getOneProcess({id: 'testProcess', version: '1', states: statesList});
-            const userWithPerimeters = new UserWithPerimeters(userMemberOfEntity1, [
-                {
-                    process: 'testProcess',
-                    state: 'responseState',
-                    rights: RightEnum.ReceiveAndWrite,
-                    filteringNotificationAllowed: true
-                }
-            ]);
-
-            const res = AcknowledgeService.isAcknowledgmentAllowed(userWithPerimeters, card, processDefinition);
-            expect(res).toBeFalse();
-        }
-    );
-
-    it(
-        'acknowledgmentAllowed of the state is OnlyWhenResponseDisabledForUser, ' +
-            'user cannot respond (user is a member of entity allowed to respond (ENTITY1) rights for the state of the response is Receive), ' +
-            'isAcknowledgmentAllowed() must return true',
-        () => {
-            statesList.set(
-                'testState',
-                new State(
-                    null,
-                    null,
-                    new Response(null, 'responseState'),
-                    AcknowledgmentAllowedEnum.ONLY_WHEN_RESPONSE_DISABLED_FOR_USER
-                )
-            );
-            const processDefinition = getOneProcess({id: 'testProcess', version: '1', states: statesList});
-            const userWithPerimeters = new UserWithPerimeters(userMemberOfEntity1, [
-                {
-                    process: 'testProcess',
-                    state: 'responseState',
-                    rights: RightEnum.Receive,
-                    filteringNotificationAllowed: true
-                }
-            ]);
-
-            const res = AcknowledgeService.isAcknowledgmentAllowed(userWithPerimeters, card, processDefinition);
-            expect(res).toBeTrue();
-        }
-    );
-
-    it(
-        'acknowledgmentAllowed of the state is OnlyWhenResponseDisabledForUser, ' +
-            'user can respond (user is a member of entity allowed to respond and user rights for the state of the response is Write), ' +
-            'lttd is not reached, isAcknowledgmentAllowed() must return false',
-        () => {
-            const lttdInTheFuture = new Date().valueOf() + 100000;
-            const cardWithLttd = getOneCard({
-                process: 'testProcess',
-                processVersion: '1',
-                state: 'testState',
-                entitiesAllowedToRespond: ['ENTITY1'],
-                lttd: lttdInTheFuture
-            });
-            statesList.set(
-                'testState',
-                new State(
-                    null,
-                    null,
-                    new Response(null, 'responseState'),
-                    AcknowledgmentAllowedEnum.ONLY_WHEN_RESPONSE_DISABLED_FOR_USER
-                )
-            );
-            const processDefinition = getOneProcess({id: 'testProcess', version: '1', states: statesList});
-            const userWithPerimeters = new UserWithPerimeters(userMemberOfEntity1, [
-                {
-                    process: 'testProcess',
-                    state: 'responseState',
-                    rights: RightEnum.ReceiveAndWrite,
-                    filteringNotificationAllowed: true
-                }
-            ]);
-
-            const res = AcknowledgeService.isAcknowledgmentAllowed(userWithPerimeters, cardWithLttd, processDefinition);
-            expect(res).toBeFalse();
-        }
-    );
-
-    it(
-        'acknowledgmentAllowed of the state is OnlyWhenResponseDisabledForUser, ' +
-            'user can respond (user is a member of entity allowed to respond and user rights for the state of the response is Write), ' +
-            'lttd is reached, isAcknowledgmentAllowed() must return true',
-        () => {
-            const lttdInThePast = new Date().valueOf() - 100000;
-            const cardWithLttd = getOneCard({
-                process: 'testProcess',
-                processVersion: '1',
-                state: 'testState',
-                entitiesAllowedToRespond: ['ENTITY1'],
-                lttd: lttdInThePast
-            });
-            statesList.set(
-                'testState',
-                new State(
-                    null,
-                    null,
-                    new Response(null, 'responseState'),
-                    AcknowledgmentAllowedEnum.ONLY_WHEN_RESPONSE_DISABLED_FOR_USER
-                )
-            );
-            const processDefinition = getOneProcess({id: 'testProcess', version: '1', states: statesList});
-            const userWithPerimeters = new UserWithPerimeters(userMemberOfEntity1, [
-                {
-                    process: 'testProcess',
-                    state: 'responseState',
-                    rights: RightEnum.ReceiveAndWrite,
-                    filteringNotificationAllowed: true
-                }
-            ]);
-
-            const res = AcknowledgeService.isAcknowledgmentAllowed(userWithPerimeters, cardWithLttd, processDefinition);
-            expect(res).toBeTrue();
-        }
-    );
-    /*****  tests with consideredAcknowledgedForUserWhen = UserHasAcknowledged  ****************************************/
-
-    it(
-        'consideredAcknowledgedForUserWhen of the state is UserHasAcknowledged, ' +
-            'no user has ack the card, no entity has ack the card, ' +
-            'isLightCardHasBeenAcknowledgedByUserOrByUserEntity() must return false',
-        () => {
-            const cardWithoutAcks = getOneCard({
-                process: 'testProcess',
-                processVersion: '1',
-                state: 'state1',
-                entityRecipients: ['ENTITY1', 'ENTITY2', 'ENTITY_FR'],
-                hasBeenAcknowledged: false,
-                entitiesAcks: []
-            });
-
-            const res = AcknowledgeService.hasLightCardBeenAcknowledgedByUserOrByUserEntity(cardWithoutAcks);
-            expect(res).toBeFalse();
-        }
-    );
-
-    it(
-        'consideredAcknowledgedForUserWhen of the state is UserHasAcknowledged, ' +
-            'user has ack the card, no other entity has ack the card, ' +
-            'isLightCardHasBeenAcknowledgedByUserOrByUserEntity() must return true',
-        () => {
-            const cardWithAck = getOneCard({
-                process: 'testProcess',
-                processVersion: '1',
-                state: 'state1',
-                entityRecipients: ['ENTITY1', 'ENTITY2', 'ENTITY_FR'],
-                hasBeenAcknowledged: true,
-                entitiesAcks: ['ENTITY1']
-            });
-
-            const res = AcknowledgeService.hasLightCardBeenAcknowledgedByUserOrByUserEntity(cardWithAck);
-            expect(res).toBeTrue();
-        }
-    );
-
-    it(
-        'consideredAcknowledgedForUserWhen of the state is UserHasAcknowledged, ' +
-            'other user (userMemberOfEntity2) has ack the card,' +
-            'isLightCardHasBeenAcknowledgedByUserOrByUserEntity() must return false',
-        () => {
-            const cardWithoutAcks = getOneCard({
-                process: 'testProcess',
-                processVersion: '1',
-                state: 'state1',
-                entityRecipients: ['ENTITY1', 'ENTITY2', 'ENTITY_FR'],
-                hasBeenAcknowledged: false,
-                entitiesAcks: ['ENTITY2']
-            });
-
-            const res = AcknowledgeService.hasLightCardBeenAcknowledgedByUserOrByUserEntity(cardWithoutAcks);
-            expect(res).toBeFalse();
-        }
-    );
-
-    it(
-        'consideredAcknowledgedForUserWhen of the state is UserHasAcknowledged, ' +
-            'other user member of ENTITY1 has ack the card,' +
-            'isLightCardHasBeenAcknowledgedByUserOrByUserEntity() must return false',
-        () => {
-            const cardWithoutAcks = getOneCard({
-                process: 'testProcess',
-                processVersion: '1',
-                state: 'state1',
-                entityRecipients: ['ENTITY1', 'ENTITY2', 'ENTITY_FR'],
-                hasBeenAcknowledged: false,
-                entitiesAcks: ['ENTITY1']
-            });
-
-            const res = AcknowledgeService.hasLightCardBeenAcknowledgedByUserOrByUserEntity(cardWithoutAcks);
-            expect(res).toBeFalse();
-        }
-    );
-    /*****  tests with consideredAcknowledgedForUserWhen = AllEntitiesOfUserHaveAcknowledged  **************************/
-
-    it(
-        'consideredAcknowledgedForUserWhen of the state is AllEntitiesOfUserHaveAcknowledged, ' +
-            'no user has ack the card, no entity has ack the card, ' +
-            'isLightCardHasBeenAcknowledgedByUserOrByUserEntity() must return false',
-        () => {
-            const cardWithoutAcks = getOneCard({
-                process: 'testProcess',
-                processVersion: '1',
-                state: 'state2',
-                entityRecipients: ['ENTITY1', 'ENTITY2', 'ENTITY_FR'],
-                hasBeenAcknowledged: false,
-                entitiesAcks: []
-            });
-
-            const res = AcknowledgeService.hasLightCardBeenAcknowledgedByUserOrByUserEntity(cardWithoutAcks);
-            expect(res).toBeFalse();
-        }
-    );
-
-    it(
-        'consideredAcknowledgedForUserWhen of the state is AllEntitiesOfUserHaveAcknowledged, ' +
-            'user has ack the card with only ENTITY1 connected, no other entity has ack the card, ' +
-            'isLightCardHasBeenAcknowledgedByUserOrByUserEntity() must return false',
-        () => {
-            usersServerMock.setResponseForCurrentUserWithPerimeter(
-                new ServerResponse(getUserMemberOfEntity1AndEntity3WithPerimeter(), ServerResponseStatus.OK, '')
-            );
-            UsersService.loadUserWithPerimetersData().subscribe();
-            const cardWithAck = getOneCard({
-                process: 'testProcess',
-                processVersion: '1',
-                state: 'state2',
-                entityRecipients: ['ENTITY1', 'ENTITY2', 'ENTITY_FR', 'ENTITY3'],
-                hasBeenAcknowledged: true,
-                entitiesAcks: ['ENTITY1']
-            });
-
-            const res = AcknowledgeService.hasLightCardBeenAcknowledgedByUserOrByUserEntity(cardWithAck);
-            expect(res).toBeFalse();
-        }
-    );
-
-    it(
-        'consideredAcknowledgedForUserWhen of the state is AllEntitiesOfUserHaveAcknowledged, ' +
-            'user has ack the card with only ENTITY1 connected, other entity has ack the card, ' +
-            'isLightCardHasBeenAcknowledgedByUserOrByUserEntity() must return false',
-        () => {
-            usersServerMock.setResponseForCurrentUserWithPerimeter(
-                new ServerResponse(getUserMemberOfEntity1AndEntity3WithPerimeter(), ServerResponseStatus.OK, '')
-            );
-            UsersService.loadUserWithPerimetersData().subscribe();
-            const cardWithAck = getOneCard({
-                process: 'testProcess',
-                processVersion: '1',
-                state: 'state2',
-                entityRecipients: ['ENTITY1', 'ENTITY2', 'ENTITY_FR', 'ENTITY3'],
-                hasBeenAcknowledged: true,
-                entitiesAcks: ['ENTITY1', 'ENTITY2']
-            });
-
-            const res = AcknowledgeService.hasLightCardBeenAcknowledgedByUserOrByUserEntity(cardWithAck);
-            expect(res).toBeFalse();
-        }
-    );
-
-    it(
-        'consideredAcknowledgedForUserWhen of the state is AllEntitiesOfUserHaveAcknowledged, ' +
-            'user has ack the card with all of its entities (ENTITY1 and ENTITY3), no other entity has ack the card, ' +
-            'isLightCardHasBeenAcknowledgedByUserOrByUserEntity() must return true',
-        () => {
-            usersServerMock.setResponseForCurrentUserWithPerimeter(
-                new ServerResponse(getUserMemberOfEntity1AndEntity3WithPerimeter(), ServerResponseStatus.OK, '')
-            );
-            UsersService.loadUserWithPerimetersData().subscribe();
-            const cardWithAck = getOneCard({
-                process: 'testProcess',
-                processVersion: '1',
-                state: 'state2',
-                entityRecipients: ['ENTITY1', 'ENTITY2', 'ENTITY_FR', 'ENTITY3'],
-                hasBeenAcknowledged: true,
-                entitiesAcks: ['ENTITY1', 'ENTITY3']
-            });
-
-            const res = AcknowledgeService.hasLightCardBeenAcknowledgedByUserOrByUserEntity(cardWithAck);
-            expect(res).toBeTrue();
-        }
-    );
-
-    it(
-        'consideredAcknowledgedForUserWhen of the state is AllEntitiesOfUserHaveAcknowledged, ' +
-            'user has not ack the card but its entities (ENTITY1 and ENTITY3) have ack the card (other members dit it), ' +
-            'isLightCardHasBeenAcknowledgedByUserOrByUserEntity() must return true',
-        () => {
-            usersServerMock.setResponseForCurrentUserWithPerimeter(
-                new ServerResponse(getUserMemberOfEntity1AndEntity3WithPerimeter(), ServerResponseStatus.OK, '')
-            );
-            UsersService.loadUserWithPerimetersData().subscribe();
-            const cardWithAck = getOneCard({
-                process: 'testProcess',
-                processVersion: '1',
-                state: 'state2',
-                entityRecipients: ['ENTITY1', 'ENTITY2', 'ENTITY_FR', 'ENTITY3'],
-                hasBeenAcknowledged: false,
-                entitiesAcks: ['ENTITY1', 'ENTITY2', 'ENTITY3']
-            });
-
-            const res = AcknowledgeService.hasLightCardBeenAcknowledgedByUserOrByUserEntity(cardWithAck);
-            expect(res).toBeTrue();
-        }
-    );
-
-    it(
-        'consideredAcknowledgedForUserWhen of the state is AllEntitiesOfUserHaveAcknowledged, ' +
-            'user has not ack the card but one of his entities (ENTITY3) have ack the card (other members dit it), ' +
-            'isLightCardHasBeenAcknowledgedByUserOrByUserEntity() must return false',
-        () => {
-            usersServerMock.setResponseForCurrentUserWithPerimeter(
-                new ServerResponse(getUserMemberOfEntity1AndEntity3WithPerimeter(), ServerResponseStatus.OK, '')
-            );
-            UsersService.loadUserWithPerimetersData().subscribe();
-            const cardWithAck = getOneCard({
-                process: 'testProcess',
-                processVersion: '1',
-                state: 'state2',
-                entityRecipients: ['ENTITY1', 'ENTITY2', 'ENTITY_FR', 'ENTITY3'],
-                hasBeenAcknowledged: false,
-                entitiesAcks: ['ENTITY2', 'ENTITY3']
-            });
-
-            const res = AcknowledgeService.hasLightCardBeenAcknowledgedByUserOrByUserEntity(cardWithAck);
-            expect(res).toBeFalse();
-        }
-    );
-
-    it(
-        'consideredAcknowledgedForUserWhen of the state is AllEntitiesOfUserHaveAcknowledged, ' +
-            'the card has no entity recipients, ' +
-            'user has not ack the card but one of his entities (ENTITY3) have ack the card (other members dit it), ' +
-            'isLightCardHasBeenAcknowledgedByUserOrByUserEntity() must return false',
-        () => {
-            usersServerMock.setResponseForCurrentUserWithPerimeter(
-                new ServerResponse(getUserMemberOfEntity1AndEntity3WithPerimeter(), ServerResponseStatus.OK, '')
-            );
-            UsersService.loadUserWithPerimetersData().subscribe();
-            const cardWithAck = getOneCard({
-                process: 'testProcess',
-                processVersion: '1',
-                state: 'state2',
-                userRecipients: ['userMemberOfEntity1', 'userMemberOfEntity2', 'userMemberOfEntity1AndEntity3'],
-                entityRecipients: [],
-                hasBeenAcknowledged: false,
-                entitiesAcks: ['ENTITY2', 'ENTITY3']
-            });
-
-            const res = AcknowledgeService.hasLightCardBeenAcknowledgedByUserOrByUserEntity(cardWithAck);
-            expect(res).toBeFalse();
-        }
-    );
-
-    it(
-        'consideredAcknowledgedForUserWhen of the state is AllEntitiesOfUserHaveAcknowledged, ' +
-            'the card has no entity recipients, ' +
-            'user has not ack the card but all of his entities (ENTITY1 and ENTITY3) have ack the card (other members dit it), ' +
-            'isLightCardHasBeenAcknowledgedByUserOrByUserEntity() must return false',
-        () => {
-            usersServerMock.setResponseForCurrentUserWithPerimeter(
-                new ServerResponse(getUserMemberOfEntity1AndEntity3WithPerimeter(), ServerResponseStatus.OK, '')
-            );
-            UsersService.loadUserWithPerimetersData().subscribe();
-            const cardWithAck = getOneCard({
-                process: 'testProcess',
-                processVersion: '1',
-                state: 'state2',
-                userRecipients: ['userMemberOfEntity1', 'userMemberOfEntity2', 'userMemberOfEntity1AndEntity3'],
-                entityRecipients: [],
-                hasBeenAcknowledged: false,
-                entitiesAcks: ['ENTITY1', 'ENTITY2', 'ENTITY3']
-            });
-
-            const res = AcknowledgeService.hasLightCardBeenAcknowledgedByUserOrByUserEntity(cardWithAck);
-            expect(res).toBeFalse();
-        }
-    );
-
-    it(
-        'consideredAcknowledgedForUserWhen of the state is AllEntitiesOfUserHaveAcknowledged, ' +
-            'the card has no entity recipients, ' +
-            'user has ack the card with all of his entities (ENTITY1 and ENTITY3), ' +
-            'isLightCardHasBeenAcknowledgedByUserOrByUserEntity() must return true',
-        () => {
-            usersServerMock.setResponseForCurrentUserWithPerimeter(
-                new ServerResponse(getUserMemberOfEntity1AndEntity3WithPerimeter(), ServerResponseStatus.OK, '')
-            );
-            UsersService.loadUserWithPerimetersData().subscribe();
-            const cardWithAck = getOneCard({
-                process: 'testProcess',
-                processVersion: '1',
-                state: 'state2',
-                userRecipients: ['userMemberOfEntity1', 'userMemberOfEntity2', 'userMemberOfEntity1AndEntity3'],
-                entityRecipients: [],
-                hasBeenAcknowledged: true,
-                entitiesAcks: ['ENTITY1', 'ENTITY3']
-            });
-
-            const res = AcknowledgeService.hasLightCardBeenAcknowledgedByUserOrByUserEntity(cardWithAck);
-            expect(res).toBeTrue();
-        }
-    );
-
-    it(
-        'consideredAcknowledgedForUserWhen of the state is AllEntitiesOfUserHaveAcknowledged, ' +
-            'the card has no entity recipients, ' +
-            'user has ack the card with only one of his entities (ENTITY3), ' +
-            'isLightCardHasBeenAcknowledgedByUserOrByUserEntity() must return true',
-        () => {
-            usersServerMock.setResponseForCurrentUserWithPerimeter(
-                new ServerResponse(getUserMemberOfEntity1AndEntity3WithPerimeter(), ServerResponseStatus.OK, '')
-            );
-            UsersService.loadUserWithPerimetersData().subscribe();
-            const cardWithAck = getOneCard({
-                process: 'testProcess',
-                processVersion: '1',
-                state: 'state2',
-                userRecipients: ['userMemberOfEntity1', 'userMemberOfEntity2', 'userMemberOfEntity1AndEntity3'],
-                entityRecipients: [],
-                hasBeenAcknowledged: true,
-                entitiesAcks: ['ENTITY3']
-            });
-
-            const res = AcknowledgeService.hasLightCardBeenAcknowledgedByUserOrByUserEntity(cardWithAck);
-            expect(res).toBeTrue();
-        }
-    );
-
-    it(
-        'consideredAcknowledgedForUserWhen of the state is AllEntitiesOfUserHaveAcknowledged, ' +
-            'the card has entity recipients but none of those of the current user, ' +
-            'user has not ack the card but one of his entities (ENTITY3) have ack the card (other members dit it), ' +
-            'isLightCardHasBeenAcknowledgedByUserOrByUserEntity() must return false',
-        () => {
-            usersServerMock.setResponseForCurrentUserWithPerimeter(
-                new ServerResponse(getUserMemberOfEntity1AndEntity3WithPerimeter(), ServerResponseStatus.OK, '')
-            );
-            UsersService.loadUserWithPerimetersData().subscribe();
-            const cardWithAck = getOneCard({
-                process: 'testProcess',
-                processVersion: '1',
-                state: 'state2',
-                userRecipients: ['userMemberOfEntity1', 'userMemberOfEntity2', 'userMemberOfEntity1AndEntity3'],
-                entityRecipients: ['ENTITY2'],
-                hasBeenAcknowledged: false,
-                entitiesAcks: ['ENTITY2', 'ENTITY3']
-            });
-
-            const res = AcknowledgeService.hasLightCardBeenAcknowledgedByUserOrByUserEntity(cardWithAck);
-            expect(res).toBeFalse();
-        }
-    );
-
-    it(
-        'consideredAcknowledgedForUserWhen of the state is AllEntitiesOfUserHaveAcknowledged, ' +
-            'the card has entity recipients but none of those of the current user, ' +
-            'user has not ack the card but all of his entities (ENTITY1 and ENTITY3) have ack the card (other members dit it), ' +
-            'isLightCardHasBeenAcknowledgedByUserOrByUserEntity() must return false',
-        () => {
-            usersServerMock.setResponseForCurrentUserWithPerimeter(
-                new ServerResponse(getUserMemberOfEntity1AndEntity3WithPerimeter(), ServerResponseStatus.OK, '')
-            );
-            UsersService.loadUserWithPerimetersData().subscribe();
-            const cardWithAck = getOneCard({
-                process: 'testProcess',
-                processVersion: '1',
-                state: 'state2',
-                userRecipients: ['userMemberOfEntity1', 'userMemberOfEntity2', 'userMemberOfEntity1AndEntity3'],
-                entityRecipients: ['ENTITY2'],
-                hasBeenAcknowledged: false,
-                entitiesAcks: ['ENTITY1', 'ENTITY2', 'ENTITY3']
-            });
-
-            const res = AcknowledgeService.hasLightCardBeenAcknowledgedByUserOrByUserEntity(cardWithAck);
-            expect(res).toBeFalse();
-        }
-    );
-
-    it(
-        'consideredAcknowledgedForUserWhen of the state is AllEntitiesOfUserHaveAcknowledged, ' +
-            'the card has entity recipients but none of those of the current user, ' +
-            'user has ack the card with all of his entities (ENTITY1 and ENTITY3), ' +
-            'isLightCardHasBeenAcknowledgedByUserOrByUserEntity() must return true',
-        () => {
-            usersServerMock.setResponseForCurrentUserWithPerimeter(
-                new ServerResponse(getUserMemberOfEntity1AndEntity3WithPerimeter(), ServerResponseStatus.OK, '')
-            );
-            UsersService.loadUserWithPerimetersData().subscribe();
-            const cardWithAck = getOneCard({
-                process: 'testProcess',
-                processVersion: '1',
-                state: 'state2',
-                userRecipients: ['userMemberOfEntity1', 'userMemberOfEntity2', 'userMemberOfEntity1AndEntity3'],
-                entityRecipients: ['ENTITY2'],
-                hasBeenAcknowledged: true,
-                entitiesAcks: ['ENTITY1', 'ENTITY3']
-            });
-
-            const res = AcknowledgeService.hasLightCardBeenAcknowledgedByUserOrByUserEntity(cardWithAck);
-            expect(res).toBeTrue();
-        }
-    );
-
-    it(
-        'consideredAcknowledgedForUserWhen of the state is AllEntitiesOfUserHaveAcknowledged, ' +
-            'the card has entity recipients but none of those of the current user, ' +
-            'user has ack the card with only one of his entities (ENTITY3), ' +
-            'isLightCardHasBeenAcknowledgedByUserOrByUserEntity() must return true',
-        () => {
-            usersServerMock.setResponseForCurrentUserWithPerimeter(
-                new ServerResponse(getUserMemberOfEntity1AndEntity3WithPerimeter(), ServerResponseStatus.OK, '')
-            );
-            UsersService.loadUserWithPerimetersData().subscribe();
-            const cardWithAck = getOneCard({
-                process: 'testProcess',
-                processVersion: '1',
-                state: 'state2',
-                userRecipients: ['userMemberOfEntity1', 'userMemberOfEntity2', 'userMemberOfEntity1AndEntity3'],
-                entityRecipients: ['ENTITY2'],
-                hasBeenAcknowledged: true,
-                entitiesAcks: ['ENTITY3']
-            });
-
-            const res = AcknowledgeService.hasLightCardBeenAcknowledgedByUserOrByUserEntity(cardWithAck);
-            expect(res).toBeTrue();
-        }
-    );
-
-    it(
-        'consideredAcknowledgedForUserWhen of the state is AllEntitiesOfUserHaveAcknowledged, ' +
-            'user is member of publisher entity ENTITY1, ' +
-            'user has ack the card with only ENTITY3 connected, other entity has ack the card, ' +
-            'isLightCardHasBeenAcknowledgedByUserOrByUserEntity() must return false',
-        () => {
-            usersServerMock.setResponseForCurrentUserWithPerimeter(
-                new ServerResponse(getUserMemberOfEntity1AndEntity3WithPerimeter(), ServerResponseStatus.OK, '')
-            );
-            UsersService.loadUserWithPerimetersData().subscribe();
-            const cardWithAck = getOneCard({
-                publisher: 'ENTITY1',
-                publisherType: 'ENTITY',
-                process: 'testProcess',
-                processVersion: '1',
-                state: 'state2',
-                entityRecipients: ['ENTITY2', 'ENTITY3'],
-                hasBeenAcknowledged: true,
-                entitiesAcks: ['ENTITY2', 'ENTITY3']
-            });
-
-            const res = AcknowledgeService.hasLightCardBeenAcknowledgedByUserOrByUserEntity(cardWithAck);
-            expect(res).toBeFalse();
-        }
-    );
-
-    it(
-        'consideredAcknowledgedForUserWhen of the state is AllEntitiesOfUserHaveAcknowledged, ' +
-            'user is member of publisher entity ENTITY1, ' +
-            'user has ack the card with all of his entities (publisher ENTITY1 and ENTITY3), other entity has ack the card, ' +
-            'isLightCardHasBeenAcknowledgedByUserOrByUserEntity() must return true',
-        () => {
-            usersServerMock.setResponseForCurrentUserWithPerimeter(
-                new ServerResponse(getUserMemberOfEntity1AndEntity3WithPerimeter(), ServerResponseStatus.OK, '')
-            );
-            UsersService.loadUserWithPerimetersData().subscribe();
-            const cardWithAck = getOneCard({
-                publisher: 'ENTITY1',
-                publisherType: 'ENTITY',
-                process: 'testProcess',
-                processVersion: '1',
-                state: 'state2',
-                entityRecipients: ['ENTITY2', 'ENTITY3'],
-                hasBeenAcknowledged: true,
-                entitiesAcks: ['ENTITY1', 'ENTITY2', 'ENTITY3']
-            });
-
-            const res = AcknowledgeService.hasLightCardBeenAcknowledgedByUserOrByUserEntity(cardWithAck);
-            expect(res).toBeTrue();
-        }
-    );
-
-    it(
-        'consideredAcknowledgedForUserWhen of the state is AllEntitiesOfUserHaveAcknowledged, ' +
-            'user has Readonly permissions, ' +
-            'user has ack the card, no other entity has ack the card, ' +
-            'isLightCardHasBeenAcknowledgedByUserOrByUserEntity() must return true',
-        () => {
-            usersServerMock.setResponseForCurrentUserWithPerimeter(
-                new ServerResponse(getUserMemberOfEntity1WithPerimeterAndReadonly(), ServerResponseStatus.OK, '')
-            );
-            UsersService.loadUserWithPerimetersData().subscribe();
-            const cardWithAck = getOneCard({
-                process: 'testProcess',
-                processVersion: '1',
-                state: 'state2',
-                entityRecipients: ['ENTITY1'],
-                hasBeenAcknowledged: true,
-                entitiesAcks: []
-            });
-
-            const res = AcknowledgeService.hasLightCardBeenAcknowledgedByUserOrByUserEntity(cardWithAck);
-            expect(res).toBeTrue();
-        }
-    );
 });
